All names in Norway

All names in Norway

Usage

nor_locations_names(border = spldata::config$border_nor)

Source

https://no.wikipedia.org/wiki/Liste_over_norske_kommunenummer

Arguments

border

The border year

Value

location_code

Location code.

location_name

Location name.

location_name_description_nb

Location name with additional description.

location_name_file_nb_utf

Location name that should be used in file names, with Norwegian characters.

location_name_file_nb_ascii

Location name that should be used in file names, without Norwegian characters.

location_order

The preferred presentation order.

granularity_geo

nation, county, municip, wardoslo, wardbergen, wardstavanger, wardtrondheim, baregion, lab.

Examples

nor_locations_names()
#>      location_code                 location_name
#>   1:         norge                         Norge
#>   2:      county42                         Agder
#>   3:      county34                     Innlandet
#>   4:      county15               Møre og Romsdal
#>   5:      county18                      Nordland
#>  ---                                            
#> 607:     lab084467    Laboratoriet Bærum sykehus
#> 608:     lab085217         Folkehelseinstituttet
#> 609:     lab095148 Lovisenberg diakonale sykehus
#> 610:     lab107883      Diakonhjemmet Sykehus AS
#> 611:     lab000012        Helgelandssykehuset HF
#>             location_name_description_nb         location_name_file_nb_utf
#>   1:                               Norge                             Norge
#>   2:                       Agder (fylke)                       Agder_fylke
#>   3:                   Innlandet (fylke)                   Innlandet_fylke
#>   4:             Møre og Romsdal (fylke)             Møre_og_Romsdal_fylke
#>   5:                    Nordland (fylke)                    Nordland_fylke
#>  ---                                                                      
#> 607:    Laboratoriet Bærum sykehus (lab)    Laboratoriet_Bærum_sykehus_lab
#> 608:         Folkehelseinstituttet (lab)         Folkehelseinstituttet_lab
#> 609: Lovisenberg diakonale sykehus (lab) Lovisenberg_diakonale_sykehus_lab
#> 610:      Diakonhjemmet Sykehus AS (lab)      Diakonhjemmet_Sykehus_AS_lab
#> 611:        Helgelandssykehuset HF (lab)        Helgelandssykehuset_HF_lab
#>            location_name_file_nb_ascii location_order granularity_geo
#>   1:                             Norge              1          nation
#>   2:                       Agder_fylke              2          county
#>   3:                   Innlandet_fylke              3          county
#>   4:             More_og_Romsdal_fylke              4          county
#>   5:                    Nordland_fylke              5          county
#>  ---                                                                 
#> 607:    Laboratoriet_Barum_sykehus_lab            607             lab
#> 608:         Folkehelseinstituttet_lab            608             lab
#> 609: Lovisenberg_diakonale_sykehus_lab            609             lab
#> 610:      Diakonhjemmet_Sykehus_AS_lab            610             lab
#> 611:        Helgelandssykehuset_HF_lab            611             lab