Population in Norway by categories

A function that easily categorizes the Norwegian population into different age categories.

Usage

nor_population_by_age_cats(
  cats = NULL,
  include_total = TRUE,
  border = spldata::config$border_nor
)

Arguments

cats

A list containing vectors that you want to categorize.

include_total

Boolean. Should 'total' be included as an age cat?

border

The year in which Norwegian geographical boundaries were designated.

Value

A data.table containing the following columns:

granularity_geo

location_code

age (as specified in the argument "cats")

sex ("total")

calyear

pop_jan1_n

imputed

Details

The current year is duplicated and added as "calyear==9999". This is in accordance with the spltidy principles regarding granularity_time=="event_*".

Examples

nor_population_by_age_cats(cats = list(c(1:10), c(11:20)))
#>        granularity_geo       location_code     age   sex calyear pop_jan1_n
#>     1:          nation               norge 001_010 total    1846     302542
#>     2:          nation               norge 001_010 total    1847     302392
#>     3:          nation               norge 001_010 total    1848     305103
#>     4:          nation               norge 001_010 total    1849     308683
#>     5:          nation               norge 001_010 total    1850     312979
#>    ---                                                                     
#> 46391:   wardstavanger wardstavanger110309   total total    9999       5221
#> 46392:   wardtrondheim wardtrondheim500101   total total    9999      52292
#> 46393:   wardtrondheim wardtrondheim500102   total total    9999      59072
#> 46394:   wardtrondheim wardtrondheim500103   total total    9999      55415
#> 46395:   wardtrondheim wardtrondheim500104   total total    9999      43414
#>        imputed
#>     1:   FALSE
#>     2:   FALSE
#>     3:   FALSE
#>     4:   FALSE
#>     5:   FALSE
#>    ---        
#> 46391:   FALSE
#> 46392:   FALSE
#> 46393:   FALSE
#> 46394:   FALSE
#> 46395:   FALSE
nor_population_by_age_cats(cats = list("one to ten" = c(1:10), "eleven to twenty" = c(11:20)))
#>        granularity_geo       location_code        age   sex calyear pop_jan1_n
#>     1:          nation               norge one to ten total    1846     302542
#>     2:          nation               norge one to ten total    1847     302392
#>     3:          nation               norge one to ten total    1848     305103
#>     4:          nation               norge one to ten total    1849     308683
#>     5:          nation               norge one to ten total    1850     312979
#>    ---                                                                        
#> 46391:   wardstavanger wardstavanger110309      total total    9999       5221
#> 46392:   wardtrondheim wardtrondheim500101      total total    9999      52292
#> 46393:   wardtrondheim wardtrondheim500102      total total    9999      59072
#> 46394:   wardtrondheim wardtrondheim500103      total total    9999      55415
#> 46395:   wardtrondheim wardtrondheim500104      total total    9999      43414
#>        imputed
#>     1:   FALSE
#>     2:   FALSE
#>     3:   FALSE
#>     4:   FALSE
#>     5:   FALSE
#>    ---        
#> 46391:   FALSE
#> 46392:   FALSE
#> 46393:   FALSE
#> 46394:   FALSE
#> 46395:   FALSE
nor_population_by_age_cats(cats = list(c(1:10), c(11:20), "021p"=c(21:200)))
#>        granularity_geo       location_code     age   sex calyear pop_jan1_n
#>     1:          nation               norge 001_010 total    1846     302542
#>     2:          nation               norge 001_010 total    1847     302392
#>     3:          nation               norge 001_010 total    1848     305103
#>     4:          nation               norge 001_010 total    1849     308683
#>     5:          nation               norge 001_010 total    1850     312979
#>    ---                                                                     
#> 61802:   wardstavanger wardstavanger110309   total total    9999       5221
#> 61803:   wardtrondheim wardtrondheim500101   total total    9999      52292
#> 61804:   wardtrondheim wardtrondheim500102   total total    9999      59072
#> 61805:   wardtrondheim wardtrondheim500103   total total    9999      55415
#> 61806:   wardtrondheim wardtrondheim500104   total total    9999      43414
#>        imputed
#>     1:   FALSE
#>     2:   FALSE
#>     3:   FALSE
#>     4:   FALSE
#>     5:   FALSE
#>    ---        
#> 61802:   FALSE
#> 61803:   FALSE
#> 61804:   FALSE
#> 61805:   FALSE
#> 61806:   FALSE